misc: accommodate requested changes
